+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.4.2
+
+### Patch Changes
+
+-   Update postcss
+-   Updated dependencies
+    -   @digitalsilk/babel-preset-default@1.4.2
+
 ## 1.4.1
 
 ### Patch Changes