+++ packages/toolkit/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.4.2
+
+### Patch Changes
+
+-   Update postcss
+-   Updated dependencies
+    -   @digitalsilk/babel-preset-default@1.4.2
+    -   @digitalsilk/stylelint-config@1.4.2
+    -   @digitalsilk/eslint-config@1.4.2
+
 ## 1.4.1
 
 ### Patch Changes